backend/requirements.txt
fastapi
uvicorn
python-multipart
langgraph
langchain-openai
python-dotenv
reportlab